# Update predict functions to align with pytorch lightning API