In response to #407, the goal of this PR is to remove any .to(CUDA) statements and let pytorch lightning manage devices. I sense that this will require updating the dataloaders for predict_image and predict_tile. I am testing locally on CPU and on HPC GPU.