@bw4sz 2023-03-23
I will also attack #404 in trying to compare the timing of this PR against predict_tile in main.