@bw4sz 2023-03-24
This passes, the same error that is unrelated in the other PR. To do items 

- [] Documentation of pytorch lightning options and integration "How do I make this faster" page
- [] Timings on cpu and gpu for a large predict tile to assess speed improvements.