@bw4sz 2023-03-24
Here are timings for a 128 mb tile, about 6000 trees, general code is 

```
from deepforest import main
import cProfile, pstats

def run():
    m = main.deepforest()
    m.config["workers"] = 0    
    print(m.config)
    m.use_release()
    boxes = m.predict_tile("2018_TEAK_3_315000_4094000_image_crop.tif")
    
    return boxes


prof = cProfile.Profile()
stats = prof.run("run()")
stats = pstats.Stats(stats)
stats.sort_stats(pstats.SortKey.TIME).print_stats(10)
```

```
(base) [b.weinstein@login3 DeepForest]$ srundev -p gpu --gpus=1 --mem 40GB --cpus-per-task 10
```
## This PR

batch size ==1
0 workers: 11.96s
1 workers:  13.86s
10 workers: 14s

batch size = 32
0 workers: 9.44s
1 workers: 9.9s
2 workers: 99s
10 workers: 11.61

## main branch

Main branch has no batch prediction, all tests are batch size =1 
0 workers: 15.572s
1 workers: 13.14s
2 workers: 13.44s
10 workers: 12.86s

The batch prediction and predict_dataloader both add time improvements. Overall its about 33% faster, and I expect those gains to be clearer at larger scales. One challenge I see, and have seen elsewhere is that workers = 0, which is no multi-processing and runs on main thread, is faster than workers > 0. 

https://discuss.pytorch.org/t/guidelines-for-assigning-num-workers-to-dataloader/813/9

I haven't played with pin_memory, its very hard to anticipate all the GPU configs that users will encounter. I think we keep workers ==1 as config default,  a slight performance hit, but safer from memory allocation.