@bw4sz 2023-04-07
Let's get this merged and then we can deal separately with the azure errors.