@ethanwhite 2023-04-07
> Let's get this merged and then we can deal separately with the azure errors.

Agreed. I'll finish looking this over as best I can in the next few minutes and then get one more review with the existing bits I've accumulated. Once you look at those and let me know we're ready I'll push the button.