@ethanwhite 2023-04-08
The new test failures are from the Pandas 2.0 release this week. I'll open a new issue.

I'll merge once the change to the loop with the aliasing is in or if you tell  me you want to keep it the way it is.