@ethanwhite 2023-04-10
Current errors seem related to #422 so merging so that we can get all of the bits and pieces in.